[UPDATE] Renaming seo.js to SEO.js
