+++ src/pages/404.js
@@ -1,6 +1,6 @@
 import React from "react"
 
-import SEO from "../components/seo"
+import SEO from "../components/SEO"
 import styled from "styled-components"
 
 const NotFoundWrapper = styled.div`
